fix: refactor classroster logic teacher dashboard
